+++ packages/ui-components/src/components/CompareTable/CompareTable.style.ts
@@ -5,20 +5,25 @@ import {
   TableRow as MuiTableRow,
   TableCell as MuiTableCell,
   tableCellClasses,
+  typographyClasses,
+  linkClasses,
   TableContainer as MuiTableContainer,
 } from "@mui/material";
 import React from "react";
 import { styled } from "../../styles";
 
-export const TableContainer = styled(MuiTableContainer)``;
+export const TableContainer = MuiTableContainer;
 
+// The table needs to have a height to allow us to set height: 100% on children
+// of TableCell.
 export const Table = styled(MuiTable)`
+  height: 100%;
   border-collapse: separate;
 `;
 
-export const TableHead = styled(MuiTableHead)``;
+export const TableHead = MuiTableHead;
 
-export const TableBody = styled(MuiTableBody)``;
+export const TableBody = MuiTableBody;
 
 // Alternate white and light grey background for table rows
 export const TableRow = styled(MuiTableRow)(({ theme }) => ({
@@ -27,6 +32,11 @@ export const TableRow = styled(MuiTableRow)(({ theme }) => ({
   },
   "&:nth-of-type(even)": {
     backgroundColor: theme.palette.common.white,
+  },
+  "&:hover": {
+    [`.${linkClasses.root} .${typographyClasses.root}`]: {
+      color: theme.palette.primary.dark,
+    },
   },
 }));