+++ packages/ui-components/src/components/MetricCompareTable/MetricCompareTable.style.ts
@@ -1,0 +1,18 @@
+import { styled } from "../../styles";
+import { Link as MuiLink } from "@mui/material";
+import { TableCell } from "../CompareTable";
+
+// We remove the padding on the table cells to ensure that the link they
+// contain covers the entire cell.
+export const StyledTableCell = styled(TableCell)`
+  padding: 0;
+`;
+
+export const StyledLink = styled(MuiLink)`
+  display: flex;
+  padding: ${({ theme }) => theme.spacing(2)};
+  text-decoration: none;
+
+  width: 100%;
+  height: 100%;
+`;